requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.4
django-cors-headers>=3.7.0
django-environ>=0.4.5
psycopg2-binary>=2.8.6
django-extensions>=3.1.3
celery>=5.0.5
redis>=3.5.3
channels>=3.0.3
channels-redis>=3.2.0
djangorestframework-simplejwt>=4.6.0
stripe>=2.55.1
Pillow>=8.1.1
